PetToyBot/requirements.txt
openai==1.10.0
langchain==0.1.4
chromadb==0.4.22
langserve==0.0.34
tiktoken==0.4.0
python-dotenv